// backend/src/types/formData.ts
export interface LongevityFormData {
  sex: "male" | "female";
  dob: string;
  country: string;
  weight: number;
  height: number;
  dietQuality: string;
  exercise: string;
  smoking: string;
  alcohol: string;
  sleepQuality: string;
  stressLevel: string;
  medicalConditions: string;
  socialConnection: string;
  incomeBracket: string;
  educationLevel: string;
  willingnessToChange: string;
}